update cleanDb.ts for build

+++ server/src/seeds/cleanDb.ts
@@ -1,11 +1,31 @@
+// import models from '../models/index.js';
+// import db from '../config/connection.js';
+
+// export default async (modelName: "Question", collectionName: string) => {
+//   try {
+//     let modelExists = await models[modelName].db.db.listCollections({
+//       name: collectionName
+//     }).toArray()
+
+//     if (modelExists.length) {
+//       await db.dropCollection(collectionName);
+//     }
+//   } catch (err) {
+//     throw err;
+//   }
+// }
 import models from '../models/index.js';
 import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    const model = models[modelName];
+
+    if (!model || !model.db?.db) {
+      throw new Error(`Database not initialized for model: ${modelName}`);
+    }
+
+    const modelExists = await model.db.db.listCollections({ name: collectionName }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
@@ -13,4 +33,4 @@ export default async (modelName: "Question", collectionName: string) => {
   } catch (err) {
     throw err;
   }
-}
+};
